third_party/WebKit/LayoutTests/editing/deleting/delete-br-011.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            'foo',
            '<div><br></div>',
            '|<br>',
        '</div>',
    ].join(''),
    'delete',
    [
        '<div contenteditable>',
            'foo',
            '<div>|<br></div>',
        '</div>',
    ].join('')),
    'Caret position after delete BR');
</script>
